- Removed outdated and buggy QTextBrowser support.     - Removed unneeded KHTMLPart support. Only QWebKit is now supported.     - Fixed zoom and "select all".     - Fixed storing scrollbar position.     - Added browser options to control JavaScript, Java and HTML5 features, which now apply to Webkit

+++ trunk/kchmviewer/src/settings.h
@@ -48,12 +48,12 @@ class Settings
 		class SavedViewWindow
 		{
 			public:
-				SavedViewWindow() { scroll_y = 0; zoom = 0; }
-				SavedViewWindow ( const QString& u, int y, int z) : url(u), scroll_y(y), zoom(z) {};
+				SavedViewWindow() { scroll_y = 0; zoom = 0.0; }
+				SavedViewWindow ( const QString& u, int y, qreal z) : url(u), scroll_y(y), zoom(z) {};
 			
 				QString		url;
 				int			scroll_y;
-				int			zoom;
+				qreal		zoom;
 		};
 		
 		typedef QList<QString>			search_saved_settings_t;
